+++ src/components/Firebase/firebase.js
@@ -1,0 +1,15 @@
+import firebase from 'firebase';
+
+// For Firebase JS SDK v7.20.0 and later, measurementId is optional
+const config = {
+  apiKey: process.env.REACT_APP_API_KEY,
+  authDomain: process.env.REACT_APP_AUTH_DOMAIN,
+  projectId: process.env.REACT_APP_PROJECT_KEY,
+  storageBucket: process.env.REACT_APP_STORAGE_BUCKET,
+  messagingSenderId: process.env.REACT_APP_MESSAGING_SENDER_ID,
+  appId: process.env.REACT_APP_APP_ID,
+  measurementId: process.env.REACT_APP_MEASUREMENT_ID
+};
+firebase.initializeApp(config);
+
+export default firebase;
