+++ src/components/Firebase/index.js
@@ -1,0 +1,1 @@
+export { default as firebase } from './firebase';
