--- src/components/index.js
@@ -1,3 +1,0 @@
-// index.js file to export all files in /components
-
-export { default as Sidenav } from './Sidenav';
